requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.10.0
pydantic-settings==2.2.1
PyJWT==2.12.1
requests==2.32.3
razorpay==1.4.1
authlib==1.3.1
groq==0.4.2
huggingface-hub==0.23.4
sqlalchemy==2.0.23
bcrypt==4.1.1
python-multipart==0.0.6
esprima==4.0.1
passlib==1.7.4